Make the longest_trip an int
Previously the longest_trip was rounded, but was still a float, resulting in
it being shown as '20.0 km' when it in reality was 19.8 km. This patch
change it to an int, so it should show up as '20 km', and will not give a
false impression of accuracy.

+++ apps/profiles/views.py
@@ -235,7 +235,7 @@ def profile(request, username, template_name="profiles/profile.html", extra_cont
             tripdataseries += '[%s, %s],' % ( nr_trips, trip.route.distance)
 
             if trip.route.distance > longest_trip:
-                longest_trip = round(trip.route.distance)
+                longest_trip = int(round(trip.route.distance))
             total_distance += trip.route.distance
 
         if trip.duration:
